Start working on a separate class for redirect logic
Let's provide a separate collaborator that cares about this (rather than
having it tied into a Session and having it have acces to everything in
the session) further, let's make sure it's super easy to test by having
it just have logic and no network access. Finally, let's type hint it to
death and test it similarly.